requirements.txt
face_recognition
paddleocr
paddlepaddle
google-genai
pypinyin
json_repair
pandas
numpy
tqdm
Pillow
scikit-learn
jupyter
opencv-python-headless

